feat(opentelemetry): Widen peer dependencies to support Otel v2 (#16246)
In preparation for supporting OpenTelemetry v2 we widened the peer
dependency range of `@sentry/opentelemetry` to allow v2 dependencies.

**Note**: `@sentry/node` still requires the v1 packages of
OpenTelemetry, this is just preparatory work to allow us to bump this
down the road.

Parts of `@sentry/opentelemetry` (and `@sentry/core`) have been reworked
to be compatible with both OpenTelemetry v1 and v2.

Unit tests in `@sentry/opentelemetry` are running with v1 dependencies
while a new package was added to `dev-packages/` with a copy of all unit
tests from `@sentry/opentelemetry` to run with v2 dependencies, adjusted
to breaking changes introduced with OpenTelemetry v2.

---------

Co-authored-by: Onur Temizkan <onur@narval.co.uk>