ci: Improve size limit tree shaking handling (#16267)
By us providing an empty terser plugin we actually get weird results
(the browser variant with tree shaking was actually larger then the
default one :O). We can just use the defaults here.

I also updated the GH action to use `bytes-iec` instead of `bytes`
package for rendering the sizes - this is also what size-limit itself
uses, so this should be better aligned now. Values may _look_ larger now
because it formats as `kB` now (so factors of 1000) instead of `KiB`
(factors of 1024).