fix(browser): Ensure pageload & navigation spans have correct data (#16279)
This PR fixes two things about our idle spans emitted from
`browserTracingIntegration`:

1. The navigation name was sometimes incorrect. This is because we look
at `window.location.pathname` at the time when the `popstate` event is
emitted - but at this point, this may not be updated yet. So a
`navigation` transaction would possibly have the pathname of the
previous page as transaction name.
2. The request data is also possibly incorrect - this is set by
`HttpContext` integration at event processing time. However, at this
time the `window.location` data is also usually already of the following
navigation, so the pageload would often have wrong request data
associated to it. Now, we store this on the current scope at span
creation time to ensure it is actually correct.