fix(node): Avoid creating breadcrumbs for suppressed requests (#16285)
It was noticed that we were creating breadcrumbs for sentry-internal
requests, which is not desired. This was introduced by us moving
breadcrumb generation out of the base http/fetch instrumentation into
our own, where we did not look at `suppressTracing`. Now, when tracing
is suppressed, no fetch/http breadcrumbs will be created.