ref(browser): Streamline browser `init()` checks (#16340)
Extracted this out of
https://github.com/getsentry/sentry-javascript/pull/15307

This PR:

1. Moves things around in the `sdk.ts` file a bit, moving types & consts
to the top of the file
2. Removes an "unecessary" log around fetch not being supported: This
log has two problems. First, it is not actually logged, because we do
`logger.log()` before running `init()`, which would actually set the
logger up. So the log would never show up. Second, the log is
superfluous, if the (default) fetch transport fails to send because
fetch is not available, you get a warning anyhow for this.
3. Streamlines the browser extension check code slightly.

Closes https://github.com/getsentry/sentry-javascript/issues/16284