Revert "feat(browser): Track measure detail as span attributes" (#16348)
It seems like this change introduced a problem where accessing
`event.detail` causes a `DOMException` to be thrown (only on firefox?).
This is causing the SDK to crash if someone has tracing enabled. Until
we figure out a reproduction and a more formal fix (with some e2e
tests) let's revert the problematic PR and move forward.

Reverts getsentry/sentry-javascript#16240

ref https://github.com/getsentry/sentry-javascript/issues/16347