test(react-router): Add and fix tests for pre rendered routes (#16346)
It was confirmed that static pre-renders don't pick up Sentry trace meta
tags — this just adds a quick test.
More context:
https://linear.app/getsentry/issue/JS-392/rr7-evaluate-static-pre-rendering

P.S. Not a priority, but for some reason, I had to add a trailing / to
fetch the correct transactions and pass the tests. I’ll look into why
this is happening for transaction route names and whether this behavior
is expected in a later PR.

---------

Co-authored-by: Charly Gomez <charly.gomez@sentry.io>