feat(node): Do not add HTTP & fetch span instrumentation if tracing is disabled (#15730)
Part of https://github.com/getsentry/sentry-javascript/issues/15725,
this PR stops adding the `HttpInstrumentation` / `UndiciInstrumentation`
if tracing is not enabled.

I _think_ this should not really be breaking, as the integrations should
not do anything else, if tracing is disabled 🤔

This only skips the http instrumentation without spans on node 22+, as
only there our own diagnostics-channel based instrumentation can cover
everything we need.