feat(browser): Disable client when browser extension is detected in `init()` (#16354)
Similar to https://github.com/getsentry/sentry-javascript/pull/16353,
this changes how default options for BrowserClient are handled. Instead
of building this in `init`, we now do (some) of it in BrowserClient.

Additionally, this also adjusts what we do if we detect a browser
extension: Instead of skipping the setup, we now just disable the
client. This streamlines this a bit and also ensures that we actually
always return a `Client` from init.

It also fixes the type for `BrowserOption` to actually allow to
configure `cdnBaseUrl`, which was actually only set for the
ClientOptions, oops.

The reason for this is to streamline the `init` code, making it easier
to extend/adjust it. Right now, there is a lot going on in different
places there. By moving as much as we can (and makes sense) to
`BrowserClient` this becomes a bit easier. While playing with different
ways to handle the browser extension stuff, I ended up landing on just
disabling the SDK in this scenario.