fix(opentelemetry): Ensure `withScope` keeps span active & `_getTraceInfoFromScope` works (#16385)
Also ensure that `withScope(scope, callback)` maintains the active span
from the passed in scope.

Fixes https://github.com/getsentry/sentry-javascript/issues/16361

The problem here was that we used the `_getSpanForScope` method in core
for this, but this only actually works for the core implementation, not
the node one 😬 So I rewrote this to use more general utilities, which
should work isomorphically. While doing this, I noticed an actual bug in
the `withScope` implementation in otel, where we did not actually keep
the active span correctly.