feat(nuxt): Added support for nuxt layers (#16372)
Added ability to detect and import config files from the **active** nuxt
layer with strategy of the **most latest** layer taking precedence.
Included fallback behavior for good measure.

Rest of the config should be merged as it already is by c12 and defu. 

Don't think this PR covers the git:repo based layers but those are
generally not full supported as well in Nuxt 3 ecosystem.

Added basic tests

Resolves
[#14345](https://github.com/getsentry/sentry-javascript/issues/14345)

---------

Co-authored-by: Ivan Pesic <pesic.ivan@feg.eu>