fix(nuxt): Add `@sentry/nuxt` as external in Rollup (#16407)
The Sentry Nuxt SDK emits the Sentry config file during the Rollup
build. Users add this file to their application which includes an import
from `@sentry/nuxt` and the init function. This file is then added as an
`.mjs` file in the build output.

However, when building the application, this `.mjs` file included a
bunch of OpenTelemetry imports and not only the init code as every
dependency from `@sentry/nuxt` was traced back and bundled into the
build output.

By adding `@sentry/nuxt` as external, the `.mjs` file really only
contains the content like in the `.ts` file.

**Could** fix
https://github.com/getsentry/sentry-javascript/issues/15204

---------

Co-authored-by: Andrei Borza <andrei.borza@sentry.io>