feat(browser): Add `unregisterOriginalCallbacks` option to `browserApiErrorsIntegration` (#16412)
This patch adds an option to fix a very unfortunate, effective double
invocation of callbacks on objects inheriting from our
`EventTarget.addEventListener` instrumentation (to be found in
`browserApiErrorsIntegration`.

Affected users can switch on the new option and doing so will 
make our instrumentation of `addEventListener` also call
`removeEventListener` on for the original callback. Applied to the
example above, it will essentially restore the idempotency of
`addEventListener`. At the same time, it feels very risky to do this so
for the time being, we'll keep this (and document it as) opt-in.

Usage:

```js
Sentry.init({
  // ...
  integrations: [
    Sentry.browserApiErrorsIntegration({
      unregisterOriginalCallbacks: true,
    }),
  ],
});
```

Docs PR: https://github.com/getsentry/sentry-docs/pull/13859

fixes https://github.com/getsentry/sentry-javascript/issues/16398