fix(sveltekit): Add import attribute for node exports (#16528)
This patch changes the `node` export in the `package.json` file to include
an `import` version (similar to the other exports) to fix
https://github.com/sveltejs/kit/issues/13869 .

In SvelteKit, we're now bundling dependencies on the server that specify
SvelteKit as a dependency or peerDependency. This has caused an issue
where builds with `@sentry/sveltekit` were being bundled incorrectly.
Adding the `import` attribute fixes this so that Vite resolves to the
ESM build of Sentry.

---------

Co-authored-by: Lukas Stracke <lukas.stracke@sentry.io>