feat(ember): Stop warning for `onError` usage (#16547)
This deprecates the `ignoreEmberOnErrorWarning` option, as it is no
longer used. Previously, we logged a warnining if we detected that this
was used. But the option is now deprecated/will be removed, and since
there is no real logic attached to this we can simply remove this (as
well as the usage of the global `Ember` namespace) to avoid
deprecations.

Closes https://github.com/getsentry/sentry-javascript/issues/16504

---------

Co-authored-by: Lukas Stracke <lukas.stracke@sentry.io>