fix(browser): Ensure `suppressTracing` does not leak when async (#16545)
This changes the behavior of `suppressTracing` to be less problematic
(or, problematic in a different way).

Today, because in the browser we do not have async context isolation,
there is only a single shared top scope. Because of this, if you have
code like this:

```js
const spanPromise = suppressTracing(async () => {
  await new Promise(resolve => setTimeout(resolve, 100));
  return startInactiveSpan({ name: 'span' });
});

const span = startInactiveSpan({ name: 'span2' });
```

The span2 will also be suppressed, because `suppressTracing` forks the
scope and sets data on it to ensure spans are not recorded. This is
problematic as that will result in completely unrelated spans, e.g.
pageload/navigation spans, being suppressed while e.g. an ongoing fetch
call that is suppressed happens.

This PR changes this to instead only suppress tracing synchronously in
the browser. This obviously is also not really ideal and can lead to
things _not_ being suppressed, but it feels like the better tradeoff for
now.