feat(node): Allow to force activate `vercelAiIntegration` (#16551)
By default, the instrumentation will register span processors only when
the ai package is used. This is done to avoid overhead of span
processing for users that do not even use this package.

However, it seems that in some environments, esp. in Next.js, the
instrumentation is not added correctly, thus never running this, and not
converting spans correctly. For now, this PR adds an escape hatch to
manually opt-in to this to still get correct spans:

```js
vercelAiIntegration({ force: true })
```