fix(vue): Ensure root component render span always ends (#16488)
This PR fixes a bug discovered in
https://github.com/getsentry/sentry-javascript/pull/16486#discussion_r2128214297
where the root component span would not end correctly if
`trackComponents` was `false`.

Also added a comment to explain the purpose of the root component
`ui.vue.render` span. We might want to look into renaming or removing
this span in the future but for now, let's fix the behaviour.