ref(node): Improve span flushing (#16577)
Follow up to https://github.com/getsentry/sentry-javascript/pull/16416,
slightly moving things around and making some small improvements to
performance and logic (some not related to that PR but another
improvement we noticed when BYK looked int this):

1. Actually debounce the flushing properly, including a maxWait of
100ms. Previously, it was technically possible to debounce flushing
forever, if a span was flushed every ms. Now, at least after 100ms it
should _always_ flush.
2. Simplify overhead by avoid checking for the 5min timeout of sent
spans. As this check `isSpanAlreadySent` has been called for every
single span that ended, it is quite high impact, and meant a little bit
of additional work (although it should not necessarily run _too_ often,
but still). Instead, we now simply check for `map.has(id)` which should
be good enough for what we want to achieve, IMHO. We still clean up the
sent spans when flushing, so old stuff should still go away eventually.
3. Made stuff that is not needed as public API private on the span
exporter class. this is technically breaking but I think it is OK, this
should not be public API surface as it does not need to be called from
outside.

For this I moved the already existing `debounce` function from replay to
core. We re-implement this in replay with a different setTimeout impl.
which is needed for some angular stuff.

---------

Co-authored-by: Abhijeet Prasad <aprasad@sentry.io>