feat(pino): Add initial package for `@sentry/pino-transport` (#16652)
ref https://github.com/getsentry/sentry-javascript/issues/15952

resolves https://github.com/getsentry/sentry-javascript/issues/16621

We've gotten a lot of feedback about supporting pino. This PR kicks that
off by creating a new pino package, specifically for the pino transport.

All pino transports need to use
https://github.com/pinojs/pino-abstract-transport, which we don't want
to add as a dep to `@sentry/node`. Hopefully keeping the versions in
sync shouldn't be too bad, although I'll force a peer dep range.