deps: Update all bundler plugin instances to latest & allow caret ranges (#16641)
If a version is released with a buggy bundler plugin dependency pinned,
it can lead to problems. IMHO it should be fine to allow `^` ranges
there to ensure bugfixes etc. are pulled in? Or is there a good reason
not to do that here? API should be stable in this regard.

Note: Astro& solidstart remain on `@sentry/vite-plugin` v2.x as that has
more stuff exposed, did not want to get into that right now... something
for v10?