feat(cloudflare): Add `instrumentWorkflowWithSentry` to instrument workflows (#16672)
- Closes #16458 

It was tricky to instrument Cloudflare workflows! The code for a
workflow might look simple but there is a lot of clever shenanigans
going on under the hood in the runtime to allow suspension of workflows.
Workflows can be hibernated at any time and all state/context inside the
your workflow class and elsewhere is lost. Ideally we want all of our
step runs to have the same `trace_id` so all steps in a workflow run are
linked together and all steps should have the same sampling decision.

To work around the state limitations, we use the workflow `instanceId`
as both the Sentry `trace_id` and the last 4 characters are used to
generate the `sample_rand` used in the sampling decision. Cloudflare
uses uuid's by default for `instanceId` but users do have the option of
passing their own IDs. If users are supplying their own `instanceId`'s,
they need to be both random and a 32 character uuid (with or without
hyphens) or the Sentry instrumentation will throw an error.

Points worthy of note:
- We use a `enableDedupe` config option (docs hidden) which removes the
`dedupeIntegration` for workflows. We want to get duplicate errors for
step retries
- We have to wrap the Cloudflare `WorkflowStep` object in another class.
The Cloudflare step object is native so it's properties can't be
overridden or proxied
- Our wrapping does end up in all the stack traces but should be
automatically hidden because they will be evaluated as `in_app: false`
- We don't wrap `step.sleep`, `step.sleepUntil` or `step.waitForEvent`
because code doesn't run after the Cloudflare native function returns ☹️
- Calling `setPropagationContext` directly on the isolation context
didn't work. It needed another `withScope` inside for
`setPropagationContext` to work. @mydea is that expected?
- This PR doesn't yet capture:
  - The payload supplied when the workflow run was started
  - The return results from the workflow steps

Here is an example trace showing the final step failing (throwing) 6
times before completing successfully. The exponential retry backoff is
clearly visible.

<img width="1233" alt="image"
src="https://github.com/user-attachments/assets/1c6356b4-2416-439c-a842-ef942fce68b4"
/>

---------

Co-authored-by: Abhijeet Prasad <aprasad@sentry.io>