feat(cloudflare): Allow interop with OpenTelemetry emitted spans (#16714)
While we eventually want to move the cloudflare SDK over to use OTEL
fully under the hood, this PR is an attempt for an intermediate solution
to allow us to still get access to e.g. the vercelAi integration, which
emits spans via `@opentelemetry/core`.

For this, we register a custom trace provider in the cloudflare SDK
which then just calls our own `startSpan` APIs. This should translate
spans to Sentry spans. The only downside is that it does not handle a
`context` being passed in for spans, so this will _not_ work for all
integrations and cases etc. But it should work for simple cases like the
`ai` package using `trace.startActiveSpan`.

TODO: Test this, verify this makes sense, ...