test(nuxt): Add fake module and composable to E2E test (#16731)
This adds a `modules` and `composables` folder to the Nuxt 4 E2E test.
The main purpose is to check, that the build runs through without
problems.

Additionally, the versions were updated (as this is the Nuxt 4 test and
it should use the latest versions for the compatibility mode).

Related to this:
https://github.com/getsentry/sentry-javascript/issues/15204#issuecomment-2948908130