feat: Add opt-in `vercelAiIntegration` to cloudflare & vercel-edge (#16732)
This is on top of
https://github.com/getsentry/sentry-javascript/pull/16714.

This adds the `vercelAiIntegration` to the cloudflare SDK, as well as to
the vercel-edge SDK.

I moved the critical code from node to core package, so we can reuse
this (the code to process spans).

The integration is not added by default, but needs to be added manually.
then it will "force" add the event processors etc. We cannot auto-detect
the `ai` package, sadly, because this does not work in workers 😢 so for
now it needs to be added manually to avoid overhead for users that don't
need this.

@andreiborza let's verify that this works when deployed to cloudflare,
esp. also the auto-enablement via the modules integration etc 🤔

---------

Co-authored-by: Andrei Borza <andrei.borza@sentry.io>
Co-authored-by: Andrei <168741329+andreiborza@users.noreply.github.com>