test: Avoid publishing old tarballs for E2E tests (#16784)
I noticed that locally I sometimes have old tarballs lying around, which
are all published for E2E tests. This is unnecessary, and can also make
it a bit unreliable which version is used for tests (as we generally
have the dependencies as `*` which could be anything). This changes it
so we only publish the current version.