fix(nuxt): Ensure order of plugins is consistent
Noticed https://github.com/getsentry/sentry-javascript/pull/16783 in the nuxt-3-min test, that the plugins are not consistently run apparently. In that test, the client integrations plugin was run before the client config plugin, leading to the client not existing yet, and thus not adding the browser tracing integration.

This PR changes this to ensure we have a consistent order of these plugins.