fix(vue): Make pageload span handling more reliable (#16799)
We used to rely on a somewhat complex heuristic to determine if a router
change is a pageload or not. This somehow did not work anymore here:
https://github.com/getsentry/sentry-javascript/pull/16783 in nuxt-3-min.
Likely some vue router difference... However, I think this can be
simplified anyhow, by just checking if we have an active pageload span.
That seems to work reliably enough.