fix(browser): Avoid 4xx response for succesful `diagnoseSdkConnectivity` request (#16840)
Our browser SDKs expose a `diagnoseSdkConnectivity` function, whose
primary purpose is to be called by the example pages injected via the
wizard. As reported in #16226, even a successful request to Sentry (i.e.
no ad blockers or other parties blocked the request) gets logged to the
console because the previous URL included an invalid public key (as well
as org id). This did not cause a `fetch` error but just a 400 (bad
request) response. However, browsers log 4xx fetch requests to the
console and users mistook this console log for an incorrectly configured
SDK setup.

This patch now changes the key to a valid but deactivated public key in a
new project in the `sentry-sdks` org. All of the information in this URL
is public and as safe to share as any DSN. So from this perspective, I
don't see concerns (happy to reconsider if reviewers have concerns). It
seems like we return 200 even for deactivated keys, which means that
now, there's no console log anymore for successful requests.

closes #16226