fix(cloudflare): calculate retries not attempts (#16834)
The `message.attempts` property starts at 1 so we were calculating the
number of attempts not the number of retries. Attempt 2 is the first
retry not attempt 1.

Currently the retry count is the same as the number of messages in the
batch this way:

![Screenshot 2025-07-07 at 23 40
48](https://github.com/user-attachments/assets/69ee5947-4677-4bd4-b0e1-33c1b62c0a10)

Additionally this causes the error rate to be quite nice and I have been
assured my code is not THAT bad:

![Screenshot 2025-07-07 at 23 41
43](https://github.com/user-attachments/assets/738cbd9b-f4fe-4213-a40c-820fe402befc)

See:
https://developers.cloudflare.com/queues/configuration/javascript-apis/