feat(vercel-edge): Do not vendor in all OpenTelemetry dependencies (#16841)
Previously, most of the OTEL dependencies have been `devDependencies` of
the vercel-edge package. This PR fixes this by properly defining the
dependencies we have.

NOTE: This was not per se incorrect, just not really necessary - we used
to inline all of these dependencies into the vercel-edge SDK.

Only the `@opentelemetry/sdk-trace-base` dependency remains a
devDependency, which means it is bundled in. Without this, edge routes
somehow fail on next 13 :/