feat(cloudflare): Add option to opt out of capturing errors in `wrapRequestHandler` (#16852)
Add a `captureErrors` option to the `wrapRequestHandler`
options from `@sentry/cloudflare`. While by default, the wrapper should
capture exceptions (for example, when it's used in pure CF worker
functions), it shouldn't do that when it's used in higher-level SDKs,
like SvelteKit. There, we want our other error capturing implementations
to do the job.