fix(sveltekit): Avoid capturing `redirect()` calls as errors in Cloudflare (#16853)
We want to avoid capturing errors too early and
too generally via the `initCloudflareSentryHandle` request handler.
Internally, this handler uses a wrapper from `@sentry/cloudflare` to
capture errors but prior to #16852 it captured everything that was
thrown. This included thrown `redirect()` objects from SvelteKit which
serve as control flow mechanisms but should not be captured as errors.

This patch opts out of capturing errors in the Cloudflare wrapper. Instead,
we rely on our already existing error capturing mechanisms in SvelteKit,
which already ignore `redirect()` and a few other error classes.