feat(v9/browser): Detect redirects when emitting navigation spans
Closes https://github.com/getsentry/sentry-javascript/issues/15286

This PR adds a new option to `browserTracingIntegration`,
`detectRedirects`, which is enabled by default. If this is enabled, the
integration will try to detect if a navigation is actually a redirect
based on a simple heuristic, and in this case, will not end the ongoing
pageload/navigation, but instead let it run and create a
`navigation.redirect` zero-duration span instead.

An example trace for this would be:
https://sentry-sdks.sentry.io/explore/discover/trace/95280de69dc844448d39de7458eab527/?dataset=transactions&eventId=8a1150fd1dc846e4ac8420ccf03ad0ee&field=title&field=project&field=user.display&field=timestamp&name=All%20Errors&project=4504956726345728&query=&queryDataset=transaction-like&sort=-timestamp&source=discover&statsPeriod=5m&timestamp=1747646096&yAxis=count%28%29

![image](https://github.com/user-attachments/assets/5f30210b-5ad0-44d9-ad49-bb00354b5fe5)

Where the respective index route that triggered this has this code:

```js
setTimeout(() => {
  window.history.pushState({}, "", "/test-sub-page");
  fetch('https://example.com')
}, 100);
```

The used heuristic is:

* If the ongoing pageload/navigation was started less than 300ms ago...
* ... and no click has happened in this time...
* ... then we consider the navigation a redirect

this limit was chosen somewhat arbitrarily, open for other suggestions
too.

While this logic will not be 100% bullet proof, it should be reliable
enough and likely better than what we have today. Users can opt-out of
this logic via `browserTracingIntegration({ detectRedirects: false })`,
if needed.

---------

Co-authored-by: Sigrid Huemer <32902192+s1gr1d@users.noreply.github.com>