feat(nuxt): Parametrize SSR routes (#16843)
Adds route parametrization to SSR server routes.

The parametrized route data is gathered during build time and saved in a
virtual file (added with
[`addTemplate`](https://nuxt.com/docs/4.x/api/kit/templates#creating-a-virtual-file-for-runtime-plugin))
which can hand-over the data to be accessed during runtime.

The `nuxt-3-min` test (Nuxt 3.7) shows that the route parametrization
does not work yet with this version. From Nuxt 3.9 onwards, it works.
This is fine, as most people are on a more recent version anyways.

part of https://github.com/getsentry/sentry-javascript/issues/16684