feat(browser): Add `beforeStartNavigationSpan` lifecycle hook (#16863)
This change adds a new `beforeStartNavigationSpan` lifecycle hook that gets
triggered directly before the `startNavigationSpan` hook by the
`startBrowserTracingNavigationSpan` helper function. The reason we need
this is because we have logic for standalone CLS and LCP spans that
needs to run _before_ we start preparing any work that's executed at the
`startNavigationSpan` hook:

- `beforeStartNavigationSpan`: send standalone span with the pageload
trace id (i.e. the currently still active trace id on the propagation
context
- `startNavigationSpan`: first recycles the propagation context, then
starts the new span

This change does not yet include switching over LCP and CLS collection to the new hook (coming in a follow-up commit)