fix(browser): Ensure standalone CLS and LCP spans have traceId of pageload span (#16864)
Fix a bug in standalone web vital spans where, if they were
sent because of a navigation, they would incorrectly be added to the
`navigation` span's trace instead of the initial `pageload` span's
trace. This surfaced while dogfooding these spans on sentry and
inspecting them in the EAP traceview.

To fix this, I added a lifecycle hook in #16863 that triggers right
before we start preparing the navigation span and hence, right before we
recycle the propagation context to the new traceId. This patch now makes
use of the new hook for CLS and LCP spans and it also makes the
integration tests stricter to check for the correct trace ids.