ref(core): Rewrite `logger` to avoid side effects (#16897)
Today the logger export from core has a side effect and reads from the global.

This PR rewrites this to instead only keep the enabled flag on the
carrier. The logger itself is a "normal" method that just looks at the
carrier to see if logging is enabled or not.