ref(browser): Unify CLS/LCP recording and add recording event attribute (#16866)
Refactor our CLS and LCP tracking and recording logic. We now
have a `listenForWebVitalReportEvents` helper function which listens to
page hides and navigation span starts and ensures that the standalone
span logic is only called once per web vital. Previously, we had this
logic in the CLS and LCP function and it was quite a lot of duplicated
and identical code.