ref(core): Avoid keeping span-FF map on GLOBAL_OBJ
Instead, we can just check the required things directly on the span. This is likely a bit more expensive than checking a map, but I would not assume this operation should be happening too excessively, so IMHO this should be fine.

This is also part of https://github.com/getsentry/sentry-javascript/issues/16846