fix(core): Avoid prolonging idle span when starting standalone span (#16928)
When we we start and send a web vital
standalone span while an idlespan (e.g. pageload) was still running,
we'd restart the idle span's child span timeout (i.e. prolong its
duration potentially). Is is unintended because by definition the
standalone span should not be associated with a potentially ongoing idle
span and its tree.

This can happen, for example when users hide the page while the pageload
span is still active, causing web vitals to be reported.