docs(aws-serverless): Fix package homepage link (#16979)
Link "homepage" on https://www.npmjs.com/package/@sentry/aws-serverless
links to a 404 page, as it targets the old `serverless` URL.

Is:
https://github.com/getsentry/sentry-javascript/tree/master/packages/serverless
Should be:
https://github.com/getsentry/sentry-javascript/tree/master/packages/aws-serverless

This is the fix.

Not sure about the commit message, feel free to pick this up and adjust.

This PR is similar to the one for GCP:
https://github.com/getsentry/sentry-javascript/pull/14411