fix(react-router): Ensure that all browser spans have `source=route` (#16984)
This updates the browser tracing implementation in react router to
ensure that we always have parametrized routes.

In the previous implementation, if the parametrized route was the same
as the "initial" one, we would just do nothing and the source remains
URL. We would also not set the origin of the spans. this is not correct,
because we actually ensured this is parametrized (there are simply no
parameters in, but still!), so the source should still be `route` in
this case.