ref(node): Add `sentry.parent_span_already_sent` attribute (#16870)
This adds a new attribute to spans that we send from the Node SDK
because their parent span was already sent. This can be used by us or
customers to debug why certain things show up as transactions/root spans
in product. We could possibly also use this in the trace view to give
users helpful pointers.