fix(node-core): Apply correct SDK metadata (#17014)
While adding the SDK to the release-registry I noticed that we never
apply the correct metadata to node-core.


https://github.com/getsentry/sentry-javascript/commit/64d056a0ddc3fd04fbb2b17a7bf9d3a30fbbaac2
failing tests show the node-core SDK getting incorrect
`sentry.javascript.node` metadata (from `NodeClient`)


https://github.com/getsentry/sentry-javascript/commit/88a975aef0e09ee1b1b0c0701b83c128518b81e5
applies correct metadata to both SDKs

I was debating changing the underlying `NodeClient` to not automatically
set `sentry.javascript.node` but wasn't sure of the implications so
decided not to.