chore(test-registry): Add more descriptive error code for common error (#16790)
The error code 137 is a very common error when running E2E tests.
Starting colima only with `colima start` often results in too little
memory and it makes sense to add a more descriptive message here.

To make it work with Colima, it's better to start it with `colima start
--cpu 10 --memory 12 --network-address`